Add the ability to set the server location and port.
This makes use of the SharedPreferences.  It will refresh and listen for the preferences
but currently the activity menu doesn't refresh.  You can exit out of the app and then
come back in and the new settings will be read.  This is a first cut just to allow
people to test the application on their own systems.