Performance tweaks for Movie Browsing.
Maintain our own cache for the background images.  The ImageLoader
cache doesn't seem to be holding them very long causing long
delays.  If the image can't be found, reluctantly run as an AsyncTask
but this really makes things slow down but the UI is much more responsive
for scrolling.

There are trade offs either way, but a more responsive UI is more
critical at this point.