Issue #9: Handle cases where there are multiple Media elements.
It appears that there can be cases where multiple media elements are
available.  Such cases are rare, but happen when the movie/show are
available in different video resolution formats.  In this case
we just are defaulting to the first option.  We may need to add
a preference setting to allow the user to choose preferred format
when streaming.