Clean up Main Menu and Moview Browser construction
The code was having everything happen in onCreate and onResume.
The problem here is that onResume is always called after onStart
which calls onResume, so the screens were getting drawn twice.

Restructured code to take into account better when the Gallery
needs to be refreshed as well as any spinners in use.  This provides
a noticeable difference in how soon the screen is available to be
used by the user.