Issue #11: Implementation of Category Filters for Movie Browser
This completes the implementation of category filters for movie browsing.
Both the single level filter as well as the secondary filter are implemented.
A soon as an item is selected it either refreshes the poster gallery or
populates the secondary filter and repoluates the poster gallery with
the first entry.

Categories are all D-Pad navigable.