Fix aspect ratio on LG and Android 4.2.1.
This should fix Issue #12.  The underlying problem was the use of
LinearLayout instead of RelativeLayout.  Also a layout_weight had been
set which was causing the SurfaceView to take up the whole screen in
certain systems.   By switching the parent view to a RelativeLayout and
specifying that the SurfaceView should be centered in the parent, this
correctly no keeps the specified aspect ratios.

Tested on Vizio CoStar with Google TV 3, and Nexus 7 running Android
4.2.1.  Bug was reproducable on Android 4.2.1.