Issues #6: Set Watched/Unwatched status
Implements the ability to set watched status when a video is played with
the internal player.  Currently not implemented with the external player
support but could be implemented by calling appropriate before the
intent is launched.

Additional meta data had to be tracked for the poster in order to easily get
the id to pass to the watched and unwatched rest urls.  Videos/Episodes now
have visual indicators in the inforgraphic details to indicate if
a video/episode has been watched or not.