Issue #7: Implement progress information.
Implemented the call back to update the progress information.  Along with
this the video player now implementes the ability to Resume from the last
recorded playback position.   The progress metadata is updated every
5 seconds on the server in a background thread.

Known issue: Need to implement the ability to fetch refreshed meta data
and update just a particular selected item when video is done playing.