Issue #14: Add Category Filtering support to TV Shows.
This enhancement adds the ability to filter TV Shows by categories.
The app currently excludes the results for filtering by Episode due
to the way plex media server intermingels the results for the two.

Plus Serenity needs a little redesign work to handle the results from
episodes and available filters for episodes.