Initial cut at adding dlna discovery of PlexMedia servers.
This makes use of cling android library for discovering upnp servers. It
does correctly discover the location of the plex media servers and
the ip address for the server can be retrieved.   We may need to look at
how we want these to be activated.

Update preference screen so that it contains a place holder for future
additions for selecting from discovered servers.  Move media player
options under their own category.