Issue #34: Implement auto discovery of the plex media servers.
Cleans up and implements using the cling dlna library auto discovery
of the available plex media servers.  A preference option is populated
with available servers, and the user can choose the one they want
to talk to through the Settings menu.  They can still override and
hard code a preference in the Server entry, but going forward it is
preferable to use the Discovered Server option.

If no servers are discovered, the user can still enter the address manually.