Issue #35: Add Able Remote support for playback on TV.
This moves the preferences around, but more importantly adds initial support
for using Able Remote to play back a video on the TV.  Future versions may
interface with the Anymote protocol directly and only launch able for
video controls.  It is out of scope for Serenity to provide full video
remote controls currently.

There is some code to try and get Serenity to be a player for Able but
it currently isn't responding to the appropriate intents.