Issue #25: Add the ability to handle Filtering Episodes
Episode filtering is limited to what is provided by Plex.  So it has
the same limitations.  When a TV Show category displays only episodes
the user is taken directly to the Episode browser screen.  If Return is
pressed it takes them back to the TV Show screen where they can select
a new category.