Issue #10: Implement Search.
This implements movie search functionality.  It makes use of the Search Dialog
provided by the system and also adds a new menu item to the main screen so
that users with out a dedicated search button can access the option.

Only movies are currently supported.  Episodes may be in the future.

Supports the following:
* Voice Searching for tablets and maybe GTV 3 (untested)
* Remembers previous searches and provides them as suggestions.