Issue #63: Enable Marquee for long menu items names.
To help with people that have really long names for their movie libraries
enabled the marquee scrolling so that the menu width stays fixed, but the
full text can be seen when selected as it scrolls horizontally across the
screen. Ideally people should choose shorter names but this is a
compromise.