Add Options menu to the Main Menu and speed up main menu.
Added the Options menu to the main menu so that it is more intuitive to
get to.  As well moved the main menu background image loading to a backround
task as it can take a bit.  this seems to speed up the main menu a bit when
navigating back and forth.