Issue #109: OSD not displaying correctly with 4.3
The internal video player's on screen display was not displaying correctly
after the upgrade to version 4.3 (api 18).  There appears to be a bug with
calculating the overly size and the parent size.  The bottom of the view
is being miscalculated and cutting off items on the screen.   To work
around this we create a special view for use with API 18 and use it instead
of the regular large view.