Issue #124: Support Plex GDM protocol for server discovery
Currently we use DLNA but not everybody has it enabled, and there have
been discovery issues using it.  It also brings in Cling which is a fairly
large library and we don't currently support any servers besides plex so
it is unecessary overhead.  Will remove DLNA server discovery in favor
of the leaner GDM implementation.