Issue #124: Set discovered servers
Once servers are discovered, and a preferred server has not been already
set, choose the first server that was found and connect to it.  Users can
change this in the settings.  Show a toast message saying that it was
auto selected.

If no servers are found, display a message to the user that they need to
set it manually via ip address.