Issue #66: Add Video Queue support
Basic Video Queue support has been added to the Internal Video player.
Basically instead of passing a bunch of intent extras a LinkedQueue is
created and the VideoContentInfo implementation is added to the queue.

The video player will retrieve the item from the queue and play it.  Items
are deleted from the queue once they are retrieved.

This sets things up so that features like Continuous Play of TV Series
or adding multiple items to a Queue can occur.