Add prompt to rate the app on the Google Play store.
Set a prompt that will ask for the app to be rated.  User has the option
to never rate it, get reminded, or rate the app.  Setting is for 14 days
after initial install to give the user time to use the app, and we only
want users that are using it fairly often.   Note a user has this on
multiple devices they will get prompted at least once on each of the devices.