Issue #140: External Player preference not being remembered
On non-googletv devices the external player was always being set to true.
This change sets a first_run flag in the preferences and and checks the
value returned.  If it is the first run it will set the external player to
true, otherwise it leaves it as is and the user can change it as they see
fit.