Issue #38: Add basic any mote support
This change adds support for Google TV remote as well as the existing
support for Able Remote.   Basically these are the two main remotes for
Google TV.  In addition, this refactoring will allow for other possible
remote and second screen functionality.   Any new app that can be
supported just needs to be added to the filter list so that it becomes
available to be used.