Improve detection and focus ability for TV devices.
* Check for android.hardware.type.television
* Check for Google TV specifically

If either of these is true then hide and show the side menu items specifically.  As they
are very sensitive to focus control and different versions of Android seem to behave
differently in this area.

Basically this makes it friendlier to remote controls but still allows the app to work on
touch screen devices.