Issue #145: Some subtitles not displaying correct characters
Depending on the parent file encoding, the strings weren't being converted
correctly.  To correct this we now use the UniversalCharacterEncoding library
to help detect the file encoding, and then use the Charset built in libraries
to convert from the original encoding format to UTF-8 for display.

This done on the client side as there is no documented public api to enable
this server side, and there are reports that it isn't reliable in its
conversion anyways.