Issue #149: Add Preference for Overscan compensation
Adds a preference to compensate for Overscan on devices like the OUYA
which don't provide a built in mechanism to adjust things like a Google TV
device does.

If enabled, will adjust the main layout to compensate for the overscan of
some TVs so that all items can be displayed on the screen.