Issue #163: Default Category always displaying when switching views.
This corrects the above issue by making sure the state of the categories
is saved and restored as appropriate when the activity restarts.  On restore
the state is passed to the Category selection listener and the prior views
category is used.

This fix only addresses the Movies and not TV Shows.