Issue #67: Create MetaData database to cache YouTube Trailer Info
In order to help with performance, particularly in Grid view, it is necessary
to cache the results from prior YouTube Queries for the trailer information.
The first time the the movie is queried, the metadata information is fetched
from YouTube, all future queries will access the cached information in
the sqlite database.

This has been implemented for both Grid View and detail views.