Issue #211: Implement Smart Play for TV Shows activity
This takes the feature and extends it so that it works in Banner view as
well as in the Grid View.  Pressing play will find the first unwatched episode
and play the video.

Pressing play in either view will play the first unwatched video.  Note
that the view currently doesn't refresh, so the watched count may be off
when it returns.