Request Resized Images From Plex Server
This greatly improves the speed of image downloading.

Also, when changing the background of the TV Grid, post a delayed task for 500ms
and cancel it if the item selection changes.
This improves the experience when moving around quickly in the grid because
it only loads the BG image when you pause for over 500ms