Add a Setting Button To Setup Overscan Compensation
This new activity alows the user to adjust the overscan compensation to be used so that
it matches exactly what his screen uses.

Change all activities that used a hardcoded setup to the proper setup.

A side affect of this is that some activities layout now seem to be a bit off.

Also, I was not able to adjust the overscan for the Preferences Activity. The problem
is that I can't find the proper view to adjust for. Using getListView() doesn't work.