Fix TV Show Suggestios Title and Description
The Android TV Recommendation card is pretty sparse in what information it
displays.  So for TV Shows, we show the Series Title, and then for the summary
the Episode Title plus Season and Episode number are displayed.

This is about as much information as can be currently included.  Anything more
just gets cut off.