Issue #271: NPE when Bundle Extras is null.
Apparently there is a chance that TVCategories can be called without passing in the bundle values.   Probably need to handle bundle instance states correctly
in the activity, but let's be gaurded here and not try to process things
without the appropriate key values.

We'll still return an Empty array list so that any progress indicator is killed.