Migrate Main Menu to a Fragment for easier swapping.
This migrates the main menu to a fragment so that it can possibly be
swapped out for a RowsFragment for use with Android TV devices.  Also
allows for easier swap out to other view implementations.

In addition created necessary layouts for when the application is run on
an unsupported normal screen device like a phone.  It will display a message
on the screen instead of crashing.