Additional Adapter refactoring
Extracts the rest of the adapters from having to use any
sort of JobManager injection.  Fetching the data is the
role of the Presenters for the main Activity View
A request can come into an activity to refresh but it
delegates that to the presenter which does the rest of
the work of requesting the actual data.

This allows the presenters to be injected if necessary.