Update to latest Android Studio and Gradle
Had to fix multiple issues as the build works a bit differently.
It doesn't necessarily bring in all transitive dependencies and
due to this will cause some false positives on annotation processing
errors.  Need to check all errors that are provided and then resolve
any compilation issues that are due to missing dependencies.