feat: Support Authentication via Username and Password
Since emby media server 4.8, the only way to authenticate is with user id and password or a pin.
This commit is based off pr #459 and updates it to the latest code base.  It supports both
servers that require a password and those that don't.  In general though, all modern Emby Media
Servers will require a password.  The password is saved and reused in future logins so the
user only has to enter it one time per machine that this is installed on.

Additional work to be done to use update to use the encrypted storage for preferences.

Co-authored-by: Hesheng Zhou <zhouhesheng@gmail.com>