fix: Emby Server 4.9 requires PlaySessionId be passed.
With Emby Server 4.9, the plabyback session id must be passed so
that it can be used during the update progress.  This is a UUID for
the playback session is generated when the session playback is started.

Multiple items could be playing at the same time for a user, so a unique
id is required.