feat: Migrate to Androidx.Media3 Exoplayer.
This migrates to the Androidx.Media3 from Exoplayer v2.
Several minor chnages in setup and playback were done to reduce
stuttering and dynamically determine of Tunnelling is supported
by the device to help improve playback on lower end devices.

Also increased the buffer size for the playback and network
optimizations.